+++ changelog.md
@@ -6,8 +6,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 # [Unreleased]
 
-# [Released]
-
 ## [0.2.11] - 2026-09-22
 
 ## Added
@@ -22,10 +20,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## Fixed
 
 - Correct doc formatting (`1f28171`)
+- Correct failing "L", "R" markers tests failing in CI (``)
 
 ## Changed
 
 - Fully enclose result within the outer operator node (`ab16d00`)
+
+# [Released]
 
 ## [0.2.10] - 2026-09-21